# apps/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Feedback Flow API",
    "version": "1.7.0"
  },
  "paths": {
...
    "/api/purchases/not-refunded": {
      "get": {
        "summary": "Get non-refunded purchases",
        "description": "Returns a paginated list of non-refunded purchases for the authenticated user. Requires read permission.",
        "tags": [
          "Purchases"
        ],
        "security": [
          {
            "bearerAuth": [
              "read:api"
            ]
          },
          {
            "oauth2": [
              "read:api"
            ]
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "default": "date"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved purchases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PurchaseSummary"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized request"
          }
        }
      }
    },
...
    "/api/purchases/refunded": {
      "get": {
        "summary": "Get refunded purchases",
        "description": "Returns a paginated list of refunded purchases for the authenticated user. Requires read permission.",
        "tags": [
          "Purchases"
        ],
        "security": [
          {
            "bearerAuth": [
              "read:api"
            ]
          },
          {
            "oauth2": [
              "read:api"
            ]
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "default": "date"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved purchases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PurchaseSummary"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PurchaseSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "order": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "refunded": {
            "type": "boolean"
          },
          "amount": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "description": "Auth0 OAuth2 Authorization Code Flow",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sctg.eu.auth0.com/authorize",
            "tokenUrl": "https://sctg.eu.auth0.com/oauth/token",
            "scopes": {
              "read:api": "Read Api",
              "write:api": "Write Api",
              "admin:api": "Admin Api",
              "backup:api": "Backup Api",
              "search:api": "Search Api",
              "auth0:admin:api": "Auth0 Admin Api",
              "openid": "Openid",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "oauth2": []
    }
  ]
}